// Client/src/lib/utils/protos/lines/ogHat.ts
export interface OgHatResponse {
  isSuccess: boolean;
  message: string;
  statusCode: number;
  data: OgHatProto[];
}

export interface OgHatProto {
  id: number;
  kodu: string;
  adi: string;
  cinsi: string;
  kesit: string;
  tipi: string;
  wkb: string;
}

function createBaseOgHatResponse(): OgHatResponse {
  return { isSuccess: false, message: "", statusCode: 0, data: [] };
}

export const OgHatResponse: MessageFns<OgHatResponse> = {
  encode(message: OgHatResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.isSuccess !== false) {
      writer.uint32(8).bool(message.isSuccess);
    }
    if (message.message !== "") {
      writer.uint32(18).string(message.message);
    }
    if (message.statusCode !== 0) {
      writer.uint32(24).int32(message.statusCode);
    }
    for (const v of message.data) {
      OgHatProto.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OgHatResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOgHatResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.isSuccess = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.statusCode = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.data.push(OgHatProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): OgHatResponse {
    return {
      isSuccess: isSet(object.isSuccess) ? globalThis.Boolean(object.isSuccess) : false,
      message: isSet(object.message) ? globalThis.String(object.message) : "",
      statusCode: isSet(object.statusCode) ? globalThis.Number(object.statusCode) : 0,
      data: globalThis.Array.isArray(object?.data) ? object.data.map((e: any) => OgHatProto.fromJSON(e)) : [],
    };
  },

  toJSON(message: OgHatResponse): unknown {
    const obj: any = {};
    if (message.isSuccess !== false) {
      obj.isSuccess = message.isSuccess;
    }
    if (message.message !== "") {
      obj.message = message.message;
    }
    if (message.statusCode !== 0) {
      obj.statusCode = Math.round(message.statusCode);
    }
    if (message.data?.length) {
      obj.data = message.data.map((e) => OgHatProto.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<OgHatResponse>, I>>(base?: I): OgHatResponse {
    return OgHatResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<OgHatResponse>, I>>(object: I): OgHatResponse {
    const message = createBaseOgHatResponse();
    message.isSuccess = object.isSuccess ?? false;
    message.message = object.message ?? "";
    message.statusCode = object.statusCode ?? 0;
    message.data = object.data?.map((e) => OgHatProto.fromPartial(e)) || [];
    return message;
  },
};

function createBaseOgHatProto(): OgHatProto {
  return { id: 0, kodu: "", adi: "", cinsi: "", kesit: "", tipi: "", wkb: "" };
}

export const OgHatProto: MessageFns<OgHatProto> = {
  encode(message: OgHatProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== 0) {
      writer.uint32(8).int32(message.id);
    }
    if (message.kodu !== "") {
      writer.uint32(18).string(message.kodu);
    }
    if (message.adi !== "") {
      writer.uint32(26).string(message.adi);
    }
    if (message.cinsi !== "") {
      writer.uint32(34).string(message.cinsi);
    }
    if (message.kesit !== "") {
      writer.uint32(42).string(message.kesit);
    }
    if (message.tipi !== "") {
      writer.uint32(50).string(message.tipi);
    }
    if (message.wkb !== "") {
      writer.uint32(58).string(message.wkb);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OgHatProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOgHatProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.id = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.kodu = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.adi = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.cinsi = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.kesit = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.tipi = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.wkb = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): OgHatProto {
    return {
      id: isSet(object.id) ? globalThis.Number(object.id) : 0,
      kodu: isSet(object.kodu) ? globalThis.String(object.kodu) : "",
      adi: isSet(object.adi) ? globalThis.String(object.adi) : "",
      cinsi: isSet(object.cinsi) ? globalThis.String(object.cinsi) : "",
      kesit: isSet(object.kesit) ? globalThis.String(object.kesit) : "",
      tipi: isSet(object.tipi) ? globalThis.String(object.tipi) : "",
      wkb: isSet(object.wkb) ? globalThis.String(object.wkb) : "",
    };
  },

  toJSON(message: OgHatProto): unknown {
    const obj: any = {};
    if (message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    if (message.kodu !== "") {
      obj.kodu = message.kodu;
    }
    if (message.adi !== "") {
      obj.adi = message.adi;
    }
    if (message.cinsi !== "") {
      obj.cinsi = message.cinsi;
    }
    if (message.kesit !== "") {
      obj.kesit = message.kesit;
    }
    if (message.tipi !== "") {
      obj.tipi = message.tipi;
    }
    if (message.wkb !== "") {
      obj.wkb = message.wkb;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<OgHatProto>, I>>(base?: I): OgHatProto {
    return OgHatProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<OgHatProto>, I>>(object: I): OgHatProto {
    const message = createBaseOgHatProto();
    message.id = object.id ?? 0;
    message.kodu = object.kodu ?? "";
    message.adi = object.adi ?? "";
    message.cinsi = object.cinsi ?? "";
    message.kesit = object.kesit ?? "";
    message.tipi = object.tipi ?? "";
    message.wkb = object.wkb ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}