// Client/src/lib/utils/protos/lines/rekortman.ts
export interface RekortmanResponse {
  isSuccess: boolean;
  message: string;
  statusCode: number;
  data: RekortmanProto[];
}

export interface RekortmanProto {
  id: number;
  kodu: string;
  adi: string;
  kesit: string;
  tipi: string;
  wkb: string;
}

function createBaseRekortmanResponse(): RekortmanResponse {
  return { isSuccess: false, message: "", statusCode: 0, data: [] };
}

export const RekortmanResponse: MessageFns<RekortmanResponse> = {
  encode(message: RekortmanResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.isSuccess !== false) {
      writer.uint32(8).bool(message.isSuccess);
    }
    if (message.message !== "") {
      writer.uint32(18).string(message.message);
    }
    if (message.statusCode !== 0) {
      writer.uint32(24).int32(message.statusCode);
    }
    for (const v of message.data) {
      RekortmanProto.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RekortmanResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRekortmanResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.isSuccess = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.statusCode = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.data.push(RekortmanProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RekortmanResponse {
    return {
      isSuccess: isSet(object.isSuccess) ? globalThis.Boolean(object.isSuccess) : false,
      message: isSet(object.message) ? globalThis.String(object.message) : "",
      statusCode: isSet(object.statusCode) ? globalThis.Number(object.statusCode) : 0,
      data: globalThis.Array.isArray(object?.data) ? object.data.map((e: any) => RekortmanProto.fromJSON(e)) : [],
    };
  },

  toJSON(message: RekortmanResponse): unknown {
    const obj: any = {};
    if (message.isSuccess !== false) {
      obj.isSuccess = message.isSuccess;
    }
    if (message.message !== "") {
      obj.message = message.message;
    }
    if (message.statusCode !== 0) {
      obj.statusCode = Math.round(message.statusCode);
    }
    if (message.data?.length) {
      obj.data = message.data.map((e) => RekortmanProto.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RekortmanResponse>, I>>(base?: I): RekortmanResponse {
    return RekortmanResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RekortmanResponse>, I>>(object: I): RekortmanResponse {
    const message = createBaseRekortmanResponse();
    message.isSuccess = object.isSuccess ?? false;
    message.message = object.message ?? "";
    message.statusCode = object.statusCode ?? 0;
    message.data = object.data?.map((e) => RekortmanProto.fromPartial(e)) || [];
    return message;
  },
};

function createBaseRekortmanProto(): RekortmanProto {
  return { id: 0, kodu: "", adi: "", kesit: "", tipi: "", wkb: "" };
}

export const RekortmanProto: MessageFns<RekortmanProto> = {
  encode(message: RekortmanProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== 0) {
      writer.uint32(8).int32(message.id);
    }
    if (message.kodu !== "") {
      writer.uint32(18).string(message.kodu);
    }
    if (message.adi !== "") {
      writer.uint32(26).string(message.adi);
    }
    if (message.kesit !== "") {
      writer.uint32(34).string(message.kesit);
    }
    if (message.tipi !== "") {
      writer.uint32(42).string(message.tipi);
    }
    if (message.wkb !== "") {
      writer.uint32(50).string(message.wkb);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RekortmanProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRekortmanProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.id = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.kodu = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.adi = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.kesit = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.tipi = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.wkb = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RekortmanProto {
    return {
      id: isSet(object.id) ? globalThis.Number(object.id) : 0,
      kodu: isSet(object.kodu) ? globalThis.String(object.kodu) : "",
      adi: isSet(object.adi) ? globalThis.String(object.adi) : "",
      kesit: isSet(object.kesit) ? globalThis.String(object.kesit) : "",
      tipi: isSet(object.tipi) ? globalThis.String(object.tipi) : "",
      wkb: isSet(object.wkb) ? globalThis.String(object.wkb) : "",
    };
  },

  toJSON(message: RekortmanProto): unknown {
    const obj: any = {};
    if (message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    if (message.kodu !== "") {
      obj.kodu = message.kodu;
    }
    if (message.adi !== "") {
      obj.adi = message.adi;
    }
    if (message.kesit !== "") {
      obj.kesit = message.kesit;
    }
    if (message.tipi !== "") {
      obj.tipi = message.tipi;
    }
    if (message.wkb !== "") {
      obj.wkb = message.wkb;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RekortmanProto>, I>>(base?: I): RekortmanProto {
    return RekortmanProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RekortmanProto>, I>>(object: I): RekortmanProto {
    const message = createBaseRekortmanProto();
    message.id = object.id ?? 0;
    message.kodu = object.kodu ?? "";
    message.adi = object.adi ?? "";
    message.kesit = object.kesit ?? "";
    message.tipi = object.tipi ?? "";
    message.wkb = object.wkb ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}